organizer service build for docker
