+++ AdminService/main.cpp
@@ -8,6 +8,24 @@ ProcessRequests processing;
 int main() {
     try {
         httplib::Server server;
+        // Обработчик preflight OPTIONS запросов
+        server.Options(".*", [&](const httplib::Request& req, httplib::Response& res) {
+            res.set_header("Access-Control-Allow-Origin", "http://localhost:3000");
+            res.set_header("Access-Control-Allow-Credentials", "true");
+            res.set_header("Access-Control-Allow-Methods", "GET, POST, OPTIONS");
+            res.set_header("Access-Control-Allow-Headers", "Content-Type, Authorization");
+            res.set_header("Content-Type", "application/json");
+            res.status = 200;
+        });
+
+        // Функция для установки CORS-заголовков
+        auto set_cors_headers = [&](httplib::Response& res) {
+            res.set_header("Access-Control-Allow-Origin", "http://localhost:3000");
+            res.set_header("Access-Control-Allow-Credentials", "true");
+            res.set_header("Access-Control-Allow-Methods", "GET, POST, OPTIONS");
+            res.set_header("Access-Control-Allow-Headers", "Content-Type, Authorization");
+            res.set_header("Content-Type", "application/json");
+        };
         std::string connect = "dbname=orchestrator host=postgres user=postgres password=postgres port=5432";
         Database db(connect);
         db.initDbFromFile("src/postgres/pending_organizers.sql");
