+++ OrganizerService/CMakeLists.txt
@@ -2,48 +2,58 @@ cmake_minimum_required(VERSION 3.10)
 project(OrganizerService VERSION 1.0.0)
 
 set(CMAKE_CXX_STANDARD 17)
-# Source files
+
+# Исходные файлы остаются теми же
 set(SOURCE_FILES
-        main.cpp
-        src/FormatRegexHelper/ValidDataChecker.h
-        src/OrganizerAccountCRUD/CreateOrganizerAccount.cpp
-        src/OrganizerAccountCRUD/CreateOrganizerAccount.h
-        src/OrganizerAccountCRUD/UpdateOrganizerAccount.cpp
-        src/OrganizerAccountCRUD/UpdateOrganizerAccount.h
-        src/OrganizerAccountCRUD/GetAccountInfo.cpp
-        src/OrganizerAccountCRUD/GetAccountInfo.h
-        src/postgres/PostgresProcessing.h
-        src/postgres/PostgresProcessing.cpp
-        src/MatchCreating/AddingTickets/AddingTickets.cpp
-        src/MatchCreating/AddingTickets/AddingTickets.h
-        src/MatchCreating/CreateMatch/MatchCreator.cpp
-        src/MatchCreating/CreateMatch/MatchCreator.h
-        src/MatchCreating/UpdateMatchInfo/Updator.cpp
-        src/MatchCreating/UpdateMatchInfo/Updator.h
-        libraries/Bcrypt.cpp/src/bcrypt.cpp
-        libraries/Bcrypt.cpp/include/bcrypt.h
-        libraries/Bcrypt.cpp/src/blowfish.cpp
+    main.cpp
+    src/FormatRegexHelper/ValidDataChecker.h
+    src/OrganizerAccountCRUD/CreateOrganizerAccount.cpp
+    src/OrganizerAccountCRUD/CreateOrganizerAccount.h
+    src/OrganizerAccountCRUD/UpdateOrganizerAccount.cpp
+    src/OrganizerAccountCRUD/UpdateOrganizerAccount.h
+    src/OrganizerAccountCRUD/GetAccountInfo.cpp
+    src/OrganizerAccountCRUD/GetAccountInfo.h
+    src/postgres/PostgresProcessing.h
+    src/postgres/PostgresProcessing.cpp
+    src/MatchCreating/AddingTickets/AddingTickets.cpp
+    src/MatchCreating/AddingTickets/AddingTickets.h
+    src/MatchCreating/CreateMatch/MatchCreator.cpp
+    src/MatchCreating/CreateMatch/MatchCreator.h
+    src/MatchCreating/UpdateMatchInfo/Updator.cpp
+    src/MatchCreating/UpdateMatchInfo/Updator.h
+    libraries/Bcrypt.cpp/src/bcrypt.cpp
+    libraries/Bcrypt.cpp/include/bcrypt.h
+    libraries/Bcrypt.cpp/src/blowfish.cpp
 )
 
-include_directories("/opt/homebrew/Cellar/curl/8.12.0/include")
-link_directories("/opt/homebrew/Cellar/curl/8.12.0/lib")
-
-include_directories("/opt/homebrew/opt/libpqxx/include")
-link_directories("/opt/homebrew/opt/libpqxx/lib")
-
-# Include directories
+# Добавьте каталоги для заголовочных файлов
+include_directories(
+    src
+    src/postgres
+    /usr/local/include
+    /usr/include
+)
 include_directories("libraries/Bcrypt.cpp/include")
-include_directories("libraries/libpqxx/include")
-include_directories("/opt/homebrew/opt/openssl@3/include")
 
-# Find the libraries
+# Подключение директорий с библиотеками
+link_directories(
+    /usr/local/lib
+    /usr/lib
+)
+
+# Ищем необходимые библиотеки
 find_package(CURL REQUIRED)
-# Create executable
+find_package(OpenSSL REQUIRED)
+find_package(PostgreSQL REQUIRED)
+
+# Создаем исполняемый файл
 add_executable(OrganizerService ${SOURCE_FILES})
 
-# Link the libraries
-target_link_libraries(OrganizerService PRIVATE CURL::libcurl
-        /opt/homebrew/opt/openssl@3/lib/libssl.dylib
-        /opt/homebrew/opt/openssl@3/lib/libcrypto.dylib
-        /opt/homebrew/opt/libpqxx/lib/libpqxx.dylib
-        pqxx)
+target_link_libraries(OrganizerService
+        PRIVATE
+        pqxx
+        pq
+        CURL::libcurl
+        OpenSSL::SSL
+        OpenSSL::Crypto
+)
