+++ OrganizerService/Dockerfile
@@ -1,5 +1,7 @@
+# Используем базовый образ Ubuntu 22.04
 FROM ubuntu:22.04
 
+# Установка необходимых пакетов
 RUN apt-get update && apt-get install -y \
     build-essential \
     cmake \
@@ -7,11 +9,12 @@ RUN apt-get update && apt-get install -y \
     libcurl4-openssl-dev \
     g++ \
     wget \
-    python3 \
-    python3-pip \
+    libssl-dev \
     git \
     && rm -rf /var/lib/apt/lists/*
 
+
+# Установка libpqxx из пакетного менеджера
 RUN wget https://github.com/jtv/libpqxx/archive/refs/tags/7.7.5.tar.gz && \
     tar -xvzf 7.7.5.tar.gz && \
     cd libpqxx-7.7.5 && \
@@ -19,24 +22,27 @@ RUN wget https://github.com/jtv/libpqxx/archive/refs/tags/7.7.5.tar.gz && \
     cmake --build build --target install && \
     cd .. && rm -rf libpqxx-7.7.5 7.7.5.tar.gz
 
-    RUN git clone https://github.com/rg3/bcrypt.git /tmp/bcrypt && \
+RUN git clone https://github.com/rg3/bcrypt.git /tmp/bcrypt && \
     cd /tmp/bcrypt && \
     make && \
     cp bcrypt.a /usr/local/lib && \
     cp bcrypt.h /usr/local/include && \
     rm -rf /tmp/bcrypt
 
+# Установка рабочего каталога
 WORKDIR /app
 
+# Копирование содержимого текущего каталога в контейнер
 COPY . .
 
-RUN rm -rf build CMakeFiles CMakeCache.txt cmake_install.cmake
-
+# Сборка приложения с помощью CMake
 RUN mkdir -p build && \
     cd build && \
     cmake .. && \
     make
 
-CMD ["./build/AuthorizationService"]
+# Используем правильный исполняемый файл
+CMD ["./build/OrganizerService"]
 
-EXPOSE 8002
+# Открытие порта
+EXPOSE 8004
