--- OrganizerService/MatchCreating/AddingTickets/AddingTickets.cpp
@@ -1,6 +1,0 @@
-#include "AddingTickets.h"
-
-void AddingTickets::AddTicketsRequest(const httplib::Request& req, httplib::Response& res,
-                       Database& db) {
-
-}
