+++ OrganizerService/src/MatchCreating/AddingTickets/AddingTickets.h
@@ -1,5 +1,5 @@
-#include "../../../../libraries/httplib.h"
-#include "../../../../libraries/nlohmann/json.hpp"
+#include "../../../libraries/httplib.h"
+#include "../../../libraries/nlohmann/json.hpp"
 #include "pqxx/pqxx"
 #include "../../postgres/PostgresProcessing.h"
